Rename fieldconfig fields Small refactoring of deserialize
