+++ src/converters/deserialize.ts
@@ -1,7 +1,7 @@
 import {Type} from "../type";
 import {FIELDS_METADATA_KEY} from "../metadata.keys";
 import {FieldMetadata} from "../field/field.metadata";
-import {NoFieldsError} from "../errors/no-fields.error";
+import {NoFieldsError} from "../errors";
 
 /**
  * Convert json for type that you need with updated names
@@ -20,8 +20,15 @@ export function deserialize<T extends Object>(json: Object, modelType: Type<T>):
     }
 
     // Simple serialization
-    fields.forEach(fieldMetadata =>
-        (<any>model)[fieldMetadata.propertyName] = fieldMetadata.serializer.deserialize((<any>json)[fieldMetadata.name]));
+    fields.forEach(getDeserialization(model, json));
 
     return model;
+}
+
+function setFieldMetadataToModel(model: Object, fieldMetadata: FieldMetadata, json: Object): void {
+    (<any>model)[fieldMetadata.modelPropertyName] = fieldMetadata.serializer.deserialize((<any>json)[fieldMetadata.jsonPropertyName])
+}
+
+function getDeserialization(model: Object, json: Object): (fieldMetadata: FieldMetadata) => void {
+    return fieldMetadata => setFieldMetadataToModel(model, fieldMetadata, json);
 }
